Trial 7.2

+++ src/estimator.py
@@ -38,14 +38,14 @@ def estimator(data):
     #factor= lost_money(duration, avgDailyIncome, avgDailyIncomePopulation )
 
     #For Impact
-    impact['casesForICUByRequestedTime'] = int( impact['infectionsByRequestedTime'] * 0.05)
-    impact['casesForVentilatorsByRequestedTime'] = int(impact['infectionsByRequestedTime'] * 0.02)
-    impact['dollarsInFlight'] = int(impact['infectionsByRequestedTime'] * avgDailyIncomePopulation * avgDailyIncome * duration)
+    impact['casesForICUByRequestedTime'] = impact['infectionsByRequestedTime'] * 0.05
+    impact['casesForVentilatorsByRequestedTime'] = impact['infectionsByRequestedTime'] * 0.02
+    impact['dollarsInFlight'] = impact['infectionsByRequestedTime'] * avgDailyIncomePopulation * avgDailyIncome * duration
 
     #For Severe Impact
-    severeImpact['casesForICUByRequestedTime'] = int( severeImpact['infectionsByRequestedTime'] * 0.05)
-    severeImpact['casesForVentilatorsByRequestedTime'] = int(severeImpact['infectionsByRequestedTime'] * 0.02)
-    severeImpact['dollarsInFlight'] = int(severeImpact['infectionsByRequestedTime'] * avgDailyIncomePopulation * avgDailyIncome * duration)
+    severeImpact['casesForICUByRequestedTime'] =  severeImpact['infectionsByRequestedTime'] * 0.05
+    severeImpact['casesForVentilatorsByRequestedTime'] = severeImpact['infectionsByRequestedTime'] * 0.02
+    severeImpact['dollarsInFlight'] = severeImpact['infectionsByRequestedTime'] * avgDailyIncomePopulation * avgDailyIncome * duration
